requirements_all.txt
matplotlib==3.9.0
nltk==3.8.1
numpy==1.26.4
opencv-python==4.9.0.80
pandas==2.2.1
scikit-learn==1.5.1

### UNIQUE REQUIREMENTS ###

# Day 3 
deep_translator==1.11.4

# Day 6 
transformers==4.38.2
torch

# Day 7
joblib==1.4.0
Pillow

# Day 9
textblob>=0.17.1
python-dotenv>=1.0.1
python-twitter-v2>=0.9.2
vaderSentiment

# Day 11
kagglehub>=0.2.9
